Add frontend auth module with role-based login, registration and admin panel
