+++ drain_eye/lib/data/datasources/local_damage_model_datasource_web.dart
@@ -1,0 +1,10 @@
+import 'package:drain_eye/domain/entities/model_inference_result.dart';
+
+// Заглушка для веб-платформы — tflite_flutter не поддерживает web.
+class LocalDamageModelDataSource {
+  Future<ModelInferenceResult> runAverageProbabilities(List<String> photoPaths) async {
+    throw UnsupportedError('TFLite недоступен в веб-версии');
+  }
+
+  void dispose() {}
+}
