+++ drain_eye/lib/presentation/screens/auth/login_screen.dart
@@ -1,14 +1,23 @@
 import 'package:drain_eye/presentation/screens/auth/auth_widgets.dart';
 import 'package:drain_eye/presentation/screens/auth/forgot_password_screen.dart';
 import 'package:drain_eye/presentation/screens/auth/register_screen.dart';
+import 'package:drain_eye/presentation/screens/admin/admin_main_screen.dart';
 import 'package:drain_eye/presentation/screens/user/main_screen.dart';
 import 'package:flutter/material.dart';
 
 /// Экран входа (UC-1) — мокап без реальной авторизации.
-/// Кнопка «Войти» переходит на MainScreen.
-/// Ссылки ведут на UC-2 (регистрация) и UC-3 (восстановление пароля).
-class LoginScreen extends StatelessWidget {
+/// Переключатель ролей: Инспектор / Администратор.
+/// Кнопка «Войти» переходит на MainScreen или AdminMainScreen.
+class LoginScreen extends StatefulWidget {
   const LoginScreen({super.key});
+
+  @override
+  State<LoginScreen> createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  // 0 = Инспектор, 1 = Администратор
+  int _selectedRole = 0;
 
   @override
   Widget build(BuildContext context) {
@@ -21,6 +30,22 @@ class LoginScreen extends StatelessWidget {
             title: 'DrainEye',
             subtitle: 'Вход в систему',
           ),
+
+          // переключатель роли
+          Container(
+            decoration: BoxDecoration(
+              color: const Color(0xFFF1F5F9),
+              borderRadius: BorderRadius.circular(10),
+            ),
+            padding: const EdgeInsets.all(3),
+            child: Row(
+              children: [
+                _roleTab('Инспектор', 0),
+                _roleTab('Администратор', 1),
+              ],
+            ),
+          ),
+          const SizedBox(height: 20),
 
           // поле Email
           const AuthLabel(label: 'Email'),
@@ -52,13 +77,16 @@ class LoginScreen extends StatelessWidget {
           ),
           const SizedBox(height: 18),
 
-          // кнопка Войти → MainScreen
+          // кнопка Войти → MainScreen или AdminMainScreen
           authPrimaryButton(
             label: 'Войти',
             onPressed: () {
+              final destination = _selectedRole == 0
+                  ? const MainScreen()
+                  : const AdminMainScreen();
               Navigator.pushReplacement(
                 context,
-                MaterialPageRoute(builder: (_) => const MainScreen()),
+                MaterialPageRoute(builder: (_) => destination),
               );
             },
           ),
@@ -116,6 +144,36 @@ class LoginScreen extends StatelessWidget {
             ],
           ),
         ],
+      ),
+    );
+  }
+
+  /// Таб переключателя роли.
+  Widget _roleTab(String label, int index) {
+    final isSelected = _selectedRole == index;
+    return Expanded(
+      child: GestureDetector(
+        onTap: () => setState(() => _selectedRole = index),
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 200),
+          height: 36,
+          decoration: BoxDecoration(
+            color: isSelected ? Colors.white : Colors.transparent,
+            borderRadius: BorderRadius.circular(8),
+            boxShadow: isSelected
+                ? [BoxShadow(color: Colors.black.withOpacity(0.06), blurRadius: 4, offset: const Offset(0, 1))]
+                : null,
+          ),
+          alignment: Alignment.center,
+          child: Text(
+            label,
+            style: TextStyle(
+              fontSize: 13,
+              fontWeight: isSelected ? FontWeight.w600 : FontWeight.w500,
+              color: isSelected ? teal : gray500,
+            ),
+          ),
+        ),
       ),
     );
   }
