+++ drain_eye/lib/presentation/screens/auth/register_screen.dart
@@ -1,11 +1,20 @@
 import 'package:drain_eye/presentation/screens/auth/auth_widgets.dart';
+import 'package:drain_eye/presentation/screens/admin/admin_main_screen.dart';
 import 'package:drain_eye/presentation/screens/user/main_screen.dart';
 import 'package:flutter/material.dart';
 
 /// Экран регистрации (UC-2) — мокап.
-/// Кнопка «Создать аккаунт» переходит на MainScreen.
-class RegisterScreen extends StatelessWidget {
+/// Выбор роли: Инспектор / Администратор.
+/// Кнопка «Создать аккаунт» переходит на MainScreen или AdminMainScreen.
+class RegisterScreen extends StatefulWidget {
   const RegisterScreen({super.key});
+
+  @override
+  State<RegisterScreen> createState() => _RegisterScreenState();
+}
+
+class _RegisterScreenState extends State<RegisterScreen> {
+  String _selectedRole = 'inspector';
 
   @override
   Widget build(BuildContext context) {
@@ -20,15 +29,44 @@ class RegisterScreen extends StatelessWidget {
           ),
 
           // поле Имя
-          const AuthLabel(label: 'Имя'),
+          const AuthLabel(label: 'ФИО'),
           const SizedBox(height: 4),
-          authInputField(placeholder: 'Иван Петров'),
+          authInputField(placeholder: 'Иванов Иван Иванович'),
           const SizedBox(height: 16),
 
           // поле Email
           const AuthLabel(label: 'Email'),
           const SizedBox(height: 4),
           authInputField(placeholder: 'user@example.com', keyboardType: TextInputType.emailAddress),
+          const SizedBox(height: 16),
+
+          // выбор роли
+          const AuthLabel(label: 'Роль'),
+          const SizedBox(height: 4),
+          Container(
+            width: double.infinity,
+            height: 44,
+            padding: const EdgeInsets.symmetric(horizontal: 14),
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(10),
+              border: Border.all(color: borderColor),
+            ),
+            child: DropdownButtonHideUnderline(
+              child: DropdownButton<String>(
+                value: _selectedRole,
+                isExpanded: true,
+                icon: const Icon(Icons.keyboard_arrow_down, color: gray400),
+                style: const TextStyle(fontSize: 14, color: gray700),
+                items: const [
+                  DropdownMenuItem(value: 'inspector', child: Text('Инспектор')),
+                  DropdownMenuItem(value: 'admin', child: Text('Администратор')),
+                ],
+                onChanged: (value) {
+                  if (value != null) setState(() => _selectedRole = value);
+                },
+              ),
+            ),
+          ),
           const SizedBox(height: 16),
 
           // поле Пароль
@@ -47,9 +85,12 @@ class RegisterScreen extends StatelessWidget {
           authPrimaryButton(
             label: 'Создать аккаунт',
             onPressed: () {
+              final destination = _selectedRole == 'inspector'
+                  ? const MainScreen()
+                  : const AdminMainScreen();
               Navigator.pushReplacement(
                 context,
-                MaterialPageRoute(builder: (_) => const MainScreen()),
+                MaterialPageRoute(builder: (_) => destination),
               );
             },
           ),
